Don't lose overlap when moving characters down to next line.
In the case where we have some part of a word that overflows then
RepositionSprites() is called to move them to the next line but we must
maintain the same sprite overlap (-2 in the case I tested) and then we
need to make sure that this overlap is also maintained when the next
character is added to the end of these copied characters.   Yes, a lot
of fuss for something that's hardly visible but we need to be perfect.