Fix iOS Accelerometer not responding.  Modified CCAccelerometer to use the polling method explained below.
The problem is that in MonoGame the implementation for Accelerometer uses the Core Motion pulling methodology that handles motion updates at specified intervals using a Queue. This is what creates the perceived buffering effect and non reactiveness to tilt movements. Handling accelerometer data this way introduces additional overhead. In the Core Motion Management documentation provided by apple most game apps are interested only the latest sample of motion data when they render a frame so should use the polling method which provides a periodic sampling of motion data.

Fix #305