Add Mouse support to Mac OSX Desktop.
Cleanup how InitialiseInputHandling and ProcessInput methods are called.  Moved these to the platform implementations so that the platform decides how these are handled instead of using #if defs.  This allows us to keep the main CCGameView as clean as possible and also implement different platform scenarios that are only Mobile (iOS, Android, Windows Phone), only Desktop (MacOSX) or a combination of the two like Windows Universal which can be deployed on different platform devices.