[project.json] Fix VS2015 compile problems. The issue is that even when compiling non Windows Universal projects NuGet still tries to installs the .net Core 4.5 references from the project.json. This gives the following error: Microsoft.NuGet.targets(89,5): error : Couldn't find the required information in the lock file. Make sure you have .NETCore,Version=v4.5.1 mentioned in your targets.
Moving the project.json to be project specific solves this without installing any extra tools or updates.  See (NuGet issue 1102)[NuGet/Home#1102]

This is tested on a 5 day old VS 2015 RTM fresh install which still gives this error even after installing the windows tools 1.1 update.  The tools update did not fix the problems but this does.

+ Add NuGet *.lock.json files to .gitignore.