Fix Transitioning Scenes to attach Events, Actions and Schedules when the transition is finished.
In the example given in the issue the event listener that is being attached is considered an event listener that should be attached when presented.  The listener is created and attached before the event dispatcher is initialised on the base CCNode.  After the scene is presented after the transition finished it then needs to be attached.

Fixes #323